+++ CHANGELOG.md
@@ -15,4 +15,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   system (without time-zone).
 - Added interface `DateTime`.
 - Added class `LocalDateTime`. (A date-time without a time-zone in the
-  ISO-8601 calendar system, such as 2007-12-03T10:15:30.)
+  ISO-8601 calendar system, such as 2007-12-03T10:15:30).
+- Added value object `LocationId`.